Moving bits around.
