+++ SpiNNaker-allocserv/src/main/java/uk/ac/manchester/spinnaker/alloc/admin/MachineDefinitionLoader.java
@@ -16,7 +16,6 @@
  */
 package uk.ac.manchester.spinnaker.alloc.admin;
 
-import static java.lang.Integer.compare;
 import static java.lang.Integer.parseInt;
 import static java.lang.Math.max;
 import static java.util.Collections.unmodifiableList;
@@ -27,7 +26,6 @@ import static org.slf4j.LoggerFactory.getLogger;
 import static org.sqlite.SQLiteErrorCode.SQLITE_CONSTRAINT_CHECK;
 import static uk.ac.manchester.spinnaker.alloc.Constants.TRIAD_CHIP_SIZE;
 import static uk.ac.manchester.spinnaker.alloc.Constants.TRIAD_DEPTH;
-import static uk.ac.manchester.spinnaker.utils.CollectionUtils.makeEnumBackingMap;
 
 import java.io.File;
 import java.io.IOException;
@@ -38,19 +36,15 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.regex.Pattern;
 
 import javax.annotation.PostConstruct;
 import javax.validation.Valid;
 import javax.validation.ValidatorFactory;
 import javax.validation.constraints.AssertFalse;
 import javax.validation.constraints.AssertTrue;
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
 
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -58,9 +52,7 @@ import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.sqlite.SQLiteException;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
@@ -73,7 +65,6 @@ import uk.ac.manchester.spinnaker.alloc.db.DatabaseEngine.Connection;
 import uk.ac.manchester.spinnaker.alloc.db.DatabaseEngine.Update;
 import uk.ac.manchester.spinnaker.alloc.model.Direction;
 import uk.ac.manchester.spinnaker.alloc.model.IPAddress;
-import uk.ac.manchester.spinnaker.machine.ChipLocation;
 
 /**
  * Loads definitions of machines from JSON, as generated by {@code py2json}.
@@ -92,393 +83,11 @@ public class MachineDefinitionLoader extends DatabaseAwareBean {
 	 * @param string
 	 *            The string containing the number to parse.
 	 * @return The parsed number.
-	 * @throws NumberFormatExeption
+	 * @throws NumberFormatException
 	 *             If the string doesn't contain such a number.
 	 */
-	private static int parseDec(String string) throws NumberFormatException {
+	static int parseDec(String string) throws NumberFormatException {
 		return parseInt(string, DECIMAL);
-	}
-
-	/**
-	 * Triad coordinates.
-	 *
-	 * @author Donal Fellows
-	 */
-	public static final class TriadCoords // FIXME
-			implements Comparable<TriadCoords> {
-		/** X coordinate. */
-		@PositiveOrZero(message = "x coordinate must not be negative")
-		public final int x;
-
-		/** Y coordinate. */
-		@PositiveOrZero(message = "y coordinate must not be negative")
-		public final int y;
-
-		/** Z coordinate. */
-		@Min(value = 0, message = "z coordinate must not be negative")
-		@Max(value = 2, message = "z coordinate must not be more than 2")
-		public final int z;
-
-		/**
-		 * Create an instance.
-		 * @param x X coordinate.
-		 * @param y Y coordinate.
-		 * @param z Z coordinate.
-		 */
-		@JsonCreator
-		public TriadCoords(@JsonProperty("x") int x, @JsonProperty("y") int y,
-				@JsonProperty("z") int z) {
-			this.x = x;
-			this.y = y;
-			this.z = z;
-		}
-
-		private static final Pattern PATTERN =
-				Pattern.compile("^\\[x:(\\d+),y:(\\d+),z:(\\d+)\\]$");
-
-		/**
-		 * Create an instance from its serial form. The serial form (where the
-		 * numbers may vary) is:
-		 *
-		 * <pre>
-		 * [x:34,y:56,z:2]
-		 * </pre>
-		 *
-		 * @param serialForm
-		 *            The form to deserialise.
-		 * @throws IllegalArgumentException
-		 *             If the string is not in the right form.
-		 */
-		@JsonCreator
-		public TriadCoords(String serialForm) {
-			var m = PATTERN.matcher(serialForm);
-			if (!m.matches()) {
-				throw new IllegalArgumentException(
-						"bad argument: " + serialForm);
-			}
-			int idx = 0;
-			x = parseDec(m.group(++idx));
-			y = parseDec(m.group(++idx));
-			z = parseDec(m.group(++idx));
-		}
-
-		private static final int TRIAD_MAJOR_OFFSET = 8;
-
-		private static final int TRIAD_MINOR_OFFSET = 4;
-
-		ChipLocation chipLocation() {
-			int rootX = x * TRIAD_CHIP_SIZE;
-			int rootY = y * TRIAD_CHIP_SIZE;
-			switch (z) {
-			case 0:
-				break;
-			case 1:
-				rootX += TRIAD_MAJOR_OFFSET;
-				rootY += TRIAD_MINOR_OFFSET;
-				break;
-			case 2:
-				rootX += TRIAD_MINOR_OFFSET;
-				rootY += TRIAD_MAJOR_OFFSET;
-				break;
-			default:
-				throw new IllegalArgumentException("bad Z coordinate");
-			}
-			return new ChipLocation(rootX, rootY);
-		}
-
-		/**
-		 * Applies a wraparound rule in a particular direction, turning
-		 * coordinate space into something of a modular field.
-		 *
-		 * @param value
-		 *            The value to wrap.
-		 * @param limit
-		 *            The upper limit. (Lower limits are always zero.)
-		 * @return The potentially wrapped value.
-		 */
-		private static int limit(int value, int limit) {
-			if (value < 0) {
-				return value + limit;
-			} else if (value >= limit) {
-				return value - limit;
-			} else {
-				return value;
-			}
-		}
-
-		/**
-		 * Get the triad coordinate that you arrive at when you move from the
-		 * current location in the indicated direction on the given machine.
-		 * This ignores dead links and dead boards.
-		 *
-		 * @param direction
-		 *            Which way to move
-		 * @param machine
-		 *            Used to determine where wraparounds are
-		 * @return The new location
-		 */
-		TriadCoords move(Direction direction, Machine machine) {
-			var di = DirInfo.get(z, direction);
-			return new TriadCoords(limit(x + di.dx, machine.getWidth()),
-					limit(y + di.dy, machine.getHeight()), z + di.dz);
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (obj instanceof TriadCoords) {
-				var other = (TriadCoords) obj;
-				return x == other.x && y == other.y && z == other.z;
-			}
-			return false;
-		}
-
-		@Override
-		public int hashCode() {
-			return (((x << 2 + x) ^ y) << 2 + y) ^ z;
-		}
-
-		@Override
-		public String toString() {
-			return "[x:" + x + ",y:" + y + ",z:" + z + "]";
-		}
-
-		@Override
-		public int compareTo(TriadCoords other) {
-			int cmp = compare(x, other.x);
-			if (cmp != 0) {
-				return cmp;
-			}
-			cmp = compare(y, other.y);
-			if (cmp != 0) {
-				return cmp;
-			}
-			return compare(z, other.z);
-		}
-	}
-
-	/**
-	 * Frame/BMP coordinates.
-	 *
-	 * @author Donal Fellows
-	 */
-	public static final class BMPCoords // FIXME
-			implements Comparable<BMPCoords> {
-		/** Cabinet number. */
-		@PositiveOrZero(message = "cabinet number must not be negative")
-		public final int c;
-
-		/** Frame number. */
-		@PositiveOrZero(message = "frame number must not be negative")
-		public final int f;
-
-		/**
-		 * Create an instance.
-		 * @param c Cabinet number.
-		 * @param f Frame number.
-		 */
-		public BMPCoords(int c, int f) {
-			this.c = c;
-			this.f = f;
-		}
-
-		private static final Pattern PATTERN =
-				Pattern.compile("^\\[c:(\\d+),f:(\\d+)\\]$");
-
-		/**
-		 * Create an instance from its serial form. The serial form (where the
-		 * numbers may vary) is:
-		 *
-		 * <pre>
-		 * [c:34,f:12]
-		 * </pre>
-		 *
-		 * @param serialForm
-		 *            The form to deserialise.
-		 * @throws IllegalArgumentException
-		 *             If the string is not in the right form.
-		 */
-		@JsonCreator
-		public BMPCoords(String serialForm) {
-			var m = PATTERN.matcher(serialForm);
-			if (!m.matches()) {
-				throw new IllegalArgumentException(
-						"bad argument: " + serialForm);
-			}
-			int idx = 0;
-			c = parseDec(m.group(++idx));
-			f = parseDec(m.group(++idx));
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (obj instanceof BMPCoords) {
-				var other = (BMPCoords) obj;
-				return c == other.c && f == other.f;
-			}
-			return false;
-		}
-
-		@Override
-		public int hashCode() {
-			return ((c << 2 + c) ^ f) << 2 + f;
-		}
-
-		@Override
-		public String toString() {
-			return "[c:" + c + ",f:" + f + "]";
-		}
-
-		@Override
-		public int compareTo(BMPCoords other) {
-			int cmp = compare(c, other.c);
-			if (cmp != 0) {
-				return cmp;
-			}
-			return compare(f, other.f);
-		}
-	}
-
-	/**
-	 * Physical board coordinates.
-	 *
-	 * @author Donal Fellows
-	 */
-	public static final class BoardPhysicalCoords // FIXME
-			implements Comparable<BoardPhysicalCoords> {
-		/** Cabinet number. */
-		@PositiveOrZero(message = "cabinet number must not be negative")
-		public final int c;
-
-		/** Frame number. */
-		@PositiveOrZero(message = "frame number must not be negative")
-		public final int f;
-
-		/** Board number. */
-		@PositiveOrZero(message = "board number must not be negative")
-		public final int b;
-
-		/**
-		 * Create an instance.
-		 * @param c Cabinet number.
-		 * @param f Frame number.
-		 * @param b Board number.
-		 */
-		@JsonCreator
-		public BoardPhysicalCoords(@JsonProperty("c") int c,
-				@JsonProperty("f") int f, @JsonProperty("b") int b) {
-			this.c = c;
-			this.f = f;
-			this.b = b;
-		}
-
-		private static final Pattern PATTERN =
-				Pattern.compile("^\\[c:(\\d+),f:(\\d+),b:(\\d+)\\]$");
-
-		/**
-		 * Create an instance from its serial form. The serial form (where the
-		 * numbers may vary) is:
-		 *
-		 * <pre>
-		 * [c:34,f:12,b:23]
-		 * </pre>
-		 *
-		 * @param serialForm
-		 *            The form to deserialise.
-		 * @throws IllegalArgumentException
-		 *             If the string is not in the right form.
-		 */
-		@JsonCreator
-		public BoardPhysicalCoords(String serialForm) {
-			var m = PATTERN.matcher(serialForm);
-			if (!m.matches()) {
-				throw new IllegalArgumentException(
-						"bad argument: " + serialForm);
-			}
-			int idx = 0;
-			c = parseDec(m.group(++idx));
-			f = parseDec(m.group(++idx));
-			b = parseDec(m.group(++idx));
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (obj instanceof BoardPhysicalCoords) {
-				var other = (BoardPhysicalCoords) obj;
-				return c == other.c && f == other.f && b == other.b;
-			}
-			return false;
-		}
-
-		@Override
-		public int hashCode() {
-			return (((c << 2 + c) ^ f) << 2 + f) ^ b;
-		}
-
-		@Override
-		public String toString() {
-			return "[c:" + c + ",f:" + f + ",b:" + b + "]";
-		}
-
-		BMPCoords bmp() {
-			return new BMPCoords(c, f);
-		}
-
-		@Override
-		public int compareTo(BoardPhysicalCoords other) {
-			int cmp = compare(c, other.c);
-			if (cmp != 0) {
-				return cmp;
-			}
-			cmp = compare(f, other.f);
-			if (cmp != 0) {
-				return cmp;
-			}
-			return compare(b, other.b);
-		}
-	}
-
-	/**
-	 * Enumeration of links from a SpiNNaker chip, as used in the old spalloc.
-	 * <p>
-	 * Note that the numbers chosen have two useful properties:
-	 *
-	 * <ul>
-	 * <li>The integer values assigned are chosen to match the numbers used to
-	 * identify the links in the low-level software API and hardware registers.
-	 * <li>The links are ordered consecutively in anticlockwise order meaning
-	 * the opposite link is {@code (link+3)%6}.
-	 * </ul>
-	 * Note that the new Spalloc uses a different notation for link directions!
-	 *
-	 * @see Direction
-	 * @author Donal Fellows
-	 */
-	public enum Link { // FIXME
-		/** East. */
-		east(Direction.SE),
-		/** North-East. */
-		northEast(Direction.E),
-		/** North. */
-		north(Direction.N),
-		/** West. */
-		west(Direction.NW),
-		/** South-West. */
-		southWest(Direction.W),
-		/** South. */
-		south(Direction.S);
-
-		private static final Map<Direction, Link> MAP =
-				makeEnumBackingMap(values(), v -> v.d);
-
-		private final Direction d;
-
-		Link(Direction d) {
-			this.d = d;
-		}
-
-		static Link of(Direction direction) {
-			return MAP.get(direction);
-		}
 	}
 
 	/**
