+++ SpiNNaker-py2json/src/test/java/uk/ac/manchester/spinnaker/py2json/TestConvert.java
@@ -37,6 +37,10 @@ import org.json.JSONException;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
 
+import uk.ac.manchester.spinnaker.machine.board.CF;
+import uk.ac.manchester.spinnaker.machine.board.CFB;
+import uk.ac.manchester.spinnaker.machine.board.XYZ;
+
 class TestConvert {
 	/**
 	 * Shortening.
